requirements.txt
html2text==2020.1.16
python-dotenv==0.20.0
selenium==4.1.5
tweepy==4.9.0
webdriver_manager==3.5.4
